Added differentiation of user rights to edit and delete comments to articles.

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Comment;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -18,7 +19,7 @@ class CommentController extends Controller
         $comment = new Comment;
         $comment->title = $request->title;
         $comment->text = $request->text;
-        $comment->author_id = Auth::id;
+        $comment->author_id = Auth::id();
         $comment->article_id = $request->article_id;
         $comment->save();
 
@@ -27,6 +28,7 @@ class CommentController extends Controller
 
     public function edit($comment_id) {
         $comment = Comment::findOrFail($comment_id);
+        Gate::authorize('comment', $comment);
         return view('comment.edit_comment', ['comment' => $comment]);
     }
 
@@ -46,6 +48,7 @@ class CommentController extends Controller
 
     public function delete($comment_id) {
         $comment = Comment::findOrFail($comment_id);
+        Gate::authorize('comment', $comment);
         $article_id = $comment->article_id;
         $comment->delete();
         return redirect()->route('article.show', ['article' => $article_id]);
