+++ resources/views/article/one_article.blade.php
@@ -48,10 +48,12 @@
                 <p class="comment__text">{{ $comment->text }}</p>
                 <p class="comment__date">{{ $comment->created_at }}</p>
                 <p class="comment__author">{{ $comment->getAuthorName() }}</p>
-                <div class="button-box">
-                    <a href="/comment/edit/{{ $comment->id }}" class="button button_blue">Редактировать</a>
-                    <a href="/comment/delete/{{ $comment->id }}" class="button button_red">Удалить</a>
-                </div>
+                @can('comment', $comment)
+                    <div class="button-box">
+                        <a href="/comment/edit/{{ $comment->id }}" class="button button_blue">Редактировать</a>
+                        <a href="/comment/delete/{{ $comment->id }}" class="button button_red">Удалить</a>
+                    </div>
+                @endcan
             </div>
         @endforeach
     </div>
